Organize imports; fix bug when default locale is not English
git-svn-id: https://unicode.org/repos/unicodetools/trunk@746 13e8329f-0b23-4da4-9fe8-d0f6fe080806

+++ unicodetools/org/unicode/text/tools/ShowUnicodeGrowth.java
@@ -91,7 +91,7 @@ public class ShowUnicodeGrowth {
             case UCD_Types.TITLECASE_LETTER:
             case UCD_Types.MODIFIER_LETTER:
             case UCD_Types.OTHER_LETTER:
-                final byte script = ucd.getScript(i);
+                final short script = ucd.getScript(i);
                 if (script == UCD_Types.HANGUL_SCRIPT) {
                     return  Type.hangul;
                 } else if (script == UCD_Types.HAN_SCRIPT) {
